Robert_mardanov.md
To be honest the assignment did put me to a test, because I think those were just the longest four hours I have spent studying. However I do want to not that I did study new commands for computer science as the amount of reddit threads and YouTube material just kept pouring in my head. I never thought I would complete it, but I made it. And apparently the issue with the code was that I had five little issues that were almost impossible to find. I do want to say that after so many failed attempts I can probably do it blindfolded and get it to work